+++ packages/plugin-session-replay-browser/src/session-replay.ts
@@ -109,6 +109,11 @@ export class SessionReplayPlugin implements EnrichmentPlugin<BrowserClient, Brow
         captureAdoptedStyleSheets: this.options.captureAdoptedStyleSheets,
         crossOriginIframes: this.options.crossOriginIframes,
         flushIntervalConfig: this.options.flushIntervalConfig,
+        eagerFullSnapshotSend: this.options.eagerFullSnapshotSend,
+        captureFullSnapshotOnFocus: this.options.captureFullSnapshotOnFocus,
+        maxPersistedEventsSizeBytes: this.options.maxPersistedEventsSizeBytes,
+        maxSingleEventSizeBytes: this.options.maxSingleEventSizeBytes,
+        sendTimeoutMs: this.options.sendTimeoutMs,
       };
 
       await this.sessionReplay.init(config.apiKey, this.srInitOptions).promise;