+++ packages/analytics-core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.50.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-core@2.49.0...@amplitude/analytics-core@2.50.0) (2026-06-11)
+
+
+### Features
+
+* **analytics-browser, analytics-core:** support referrerPolicy in FetchTransport ([#1805](https://github.com/amplitude/Amplitude-TypeScript/issues/1805)) ([871e432](https://github.com/amplitude/Amplitude-TypeScript/commit/871e4326debc1c90028fa305056d9bbe234e0e1e))
+
+
+
+
+
 # [2.49.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-core@2.48.2...@amplitude/analytics-core@2.49.0) (2026-06-05)